+++ GendocsForms/Forms/FrmPantallaConf.cs
