+++ GendocsForms/Forms/FrmPantallaConf.resx
