+++ GendocsForms/GendocsForms.csproj
@@ -51,9 +51,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Update="Forms\FrmMantenimientoEmpleados.cs">
-      <SubType>Form</SubType>
-    </Compile>
     <Compile Update="Properties\Resources.Designer.cs">
       <DesignTime>True</DesignTime>
       <AutoGen>True</AutoGen>
