+++ WindowsFormsApp1/WindowsFormsApp1.csproj
@@ -48,14 +48,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\GendocsForms\Forms\FrmAyuda.cs">
-      <Link>FrmAyuda.cs</Link>
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="..\GendocsForms\Forms\FrmAyuda.designer.cs">
-      <Link>FrmAyuda.designer.cs</Link>
-      <DependentUpon>FrmAyuda.cs</DependentUpon>
-    </Compile>
     <Compile Include="..\GendocsForms\Forms\FrmClientes.cs">
       <Link>FrmClientes.cs</Link>
       <SubType>Form</SubType>
@@ -71,14 +63,6 @@
     <Compile Include="..\GendocsForms\Forms\FrmContactos.designer.cs">
       <Link>FrmContactos.designer.cs</Link>
       <DependentUpon>FrmContactos.cs</DependentUpon>
-    </Compile>
-    <Compile Include="..\GendocsForms\Forms\FrmContactos2.cs">
-      <Link>FrmContactos2.cs</Link>
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="..\GendocsForms\Forms\FrmContactos2.designer.cs">
-      <Link>FrmContactos2.designer.cs</Link>
-      <DependentUpon>FrmContactos2.cs</DependentUpon>
     </Compile>
     <Compile Include="..\GendocsForms\Forms\frmEmpleadosFML.cs">
       <Link>frmEmpleadosFML.cs</Link>
@@ -143,14 +127,6 @@
     <Compile Include="..\GendocsForms\Forms\FrmMantenimientoEmpleados.designer.cs">
       <Link>FrmMantenimientoEmpleados.designer.cs</Link>
       <DependentUpon>FrmMantenimientoEmpleados.cs</DependentUpon>
-    </Compile>
-    <Compile Include="..\GendocsForms\Forms\FrmMantenimientoEmpleados2.cs">
-      <Link>FrmMantenimientoEmpleados2.cs</Link>
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="..\GendocsForms\Forms\FrmMantenimientoEmpleados2.designer.cs">
-      <Link>FrmMantenimientoEmpleados2.designer.cs</Link>
-      <DependentUpon>FrmMantenimientoEmpleados2.cs</DependentUpon>
     </Compile>
     <Compile Include="..\GendocsForms\Forms\frmMantenimientoEmpleadosFML.cs">
       <Link>frmMantenimientoEmpleadosFML.cs</Link>
@@ -242,10 +218,6 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <EmbeddedResource Include="..\GendocsForms\Forms\FrmAyuda.resx">
-      <Link>FrmAyuda.resx</Link>
-      <DependentUpon>FrmAyuda.cs</DependentUpon>
-    </EmbeddedResource>
     <EmbeddedResource Include="..\GendocsForms\Forms\FrmClientes.resx">
       <Link>FrmClientes.resx</Link>
       <DependentUpon>FrmClientes.cs</DependentUpon>
@@ -253,10 +225,6 @@
     <EmbeddedResource Include="..\GendocsForms\Forms\FrmContactos.resx">
       <Link>FrmContactos.resx</Link>
       <DependentUpon>FrmContactos.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="..\GendocsForms\Forms\FrmContactos2.resx">
-      <Link>FrmContactos2.resx</Link>
-      <DependentUpon>FrmContactos2.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="..\GendocsForms\Forms\frmEmpleadosFML.resx">
       <Link>frmEmpleadosFML.resx</Link>
@@ -289,10 +257,6 @@
     <EmbeddedResource Include="..\GendocsForms\Forms\FrmMantenimientoEmpleados.resx">
       <Link>FrmMantenimientoEmpleados.resx</Link>
       <DependentUpon>FrmMantenimientoEmpleados.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="..\GendocsForms\Forms\FrmMantenimientoEmpleados2.resx">
-      <Link>FrmMantenimientoEmpleados2.resx</Link>
-      <DependentUpon>FrmMantenimientoEmpleados2.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="..\GendocsForms\Forms\frmMantenimientoEmpleadosFML.resx">
       <Link>frmMantenimientoEmpleadosFML.resx</Link>
